fix: handle deletes

+++ cognite_toolkit/_cdf_tk/cruds/_resource_cruds/function.py
@@ -22,6 +22,7 @@ from cognite.client.data_classes.capabilities import (
     FunctionsAcl,
     SessionsAcl,
 )
+from cognite.client.data_classes.data_modeling import NodeId
 from cognite.client.data_classes.data_modeling.cdm.v1 import CogniteFileApply
 from cognite.client.data_classes.functions import HANDLER_FILE_NAME
 from cognite.client.exceptions import CogniteAPIError
@@ -406,7 +407,18 @@ class FunctionCRUD(ResourceCRUD[str, FunctionWrite, Function]):
 
         self.client.functions.delete_with_429_retry(external_id=ids, ignore_unknown_ids=True)
         file_ids = {func.file_id for func in functions if func.file_id}
-        self.client.files.delete(id=list(file_ids), ignore_unknown_ids=True)
+        files = self.client.files.retrieve_multiple(list(file_ids))
+        dm_file_nodes: set[NodeId] = set()
+        file_ids = set()
+        for file in files:
+            if file.instance_id is not None:
+                dm_file_nodes.add(file.instance_id)
+            else:
+                file_ids.add(file.id)
+        if file_ids:
+            self.client.files.delete(id=list(file_ids), ignore_unknown_ids=True)
+        if dm_file_nodes:
+            self.client.data_modeling.instances.delete(list(dm_file_nodes))
         return len(ids)
 
     def _iterate(
